# Add minor performance improvements on methods using mapAsCollection