Since the size of destination collection is equal (or possibly a bit lower in *mapAsSet* and *mapAsMap* methods) to the number of elements of source iterable, the initial capacity of the destination collection can most of the times be specified on its creation.

Therefore in case the source iterable is subtype of *java.util.Collection*, its size is retrieved and used to specify initial capacity of destination collection. Otherwise destination collection is instantiated with default initial capacity as for iterables which are not subtype of *java.util.Collection* it is required iterating through each their element in order to get their amount.

**Benchmarking**
scenario: transforming iterable of 2,500,000 elements into corresponding DTO objects using Orika's mapAsList method
* only transformation time is recorded
* tests were repeated 50 times
* case 1: iterable object is subtype of Collection
  * result before changes: 2.696s (average), 2.319s (median) - resizing of internal array is performed 31 times thus causing more memory consumption
  * result after changes: 2.589s (average), 2.314s (median) - no need for additional resizing
* case 2: iterable object isn't subtype of Collection
  * no difference